Seperate components into files and Do some improvements by eslint
